Consolidate safari tests to minimize sessions
