+++ lib/commands/context.js
@@ -379,7 +379,7 @@ extensions.typeAndNavToUrl = async function () {
     // get the last address element and set the url
     // this is flakey on certain systems so we retry until we get something
     let elId = await retryInterval(5, 1000, async () => {
-      let el = _.last(await this.findElements('class name', 'UIATextField'));
+      let el = _.first(await this.findElements('class name', 'UIATextField'));
       return el.ELEMENT;
     });
     await this.setValueImmediate(this.getCurrentUrl(), elId);
