+++ test/e2e/driver/built-in-apps-e2e-specs.js
@@ -17,7 +17,6 @@ describe('built-in apps', function () {
       let caps = {
         app: 'Calendar',
         platformName: 'iOS',
-        deviceName: 'iPhone 6',
         showIOSLog: false,
         fullReset: false,
         newCommandTimeout: 120
