+++ test/e2e/safari/webview/execute-specs.js
@@ -9,50 +9,68 @@ const GET_WRONG_INNERHTML = `return document.body.innerHTML.indexOf('I am not so
 const GET_ELEM_BY_TAGNAME = `return document.getElementsByTagName('a');`;
 
 describe('safari - webview - execute', function() {
-  const driver = setup(this, desired, {noReset: true}).driver;
+  const driver = setup(this, desired, {noReset: true}, false, true).driver;
   before(async () => await loadWebView(desired, driver));
 
-  it('should bubble up javascript errors', async () => {
-    expect(async () => await driver.execute(`'nan'--`)).to.throw;
+  describe('synchronous', function () {
+    it('should bubble up javascript errors', async () => {
+      await driver.execute(`'nan'--`).should.eventually.be.rejected;
+    });
+
+    it('should eval javascript', async () => {
+      (await driver.execute('return 1')).should.be.equal(1);
+    });
+
+    it('should not be returning hardcoded results', async () => {
+      (await driver.execute('return 1+1')).should.be.equal(2);
+    });
+
+    it(`should return nothing when you don't explicitly return`, async () => {
+      expect(await driver.execute('1+1')).to.not.exist;
+    });
+
+    it('should execute code inside the web view', async () => {
+      (await driver.execute(GET_RIGHT_INNERHTML)).should.be.ok;
+      (await driver.execute(GET_WRONG_INNERHTML)).should.not.be.ok;
+    });
+
+    it('should convert selenium element arg to webview element', async () => {
+      let el = await driver.findElement('id', 'useragent');
+      await driver.execute(SCROLL_INTO_VIEW, [el]);
+    });
+
+    it('should catch stale or undefined element as arg', async () => {
+      let el = await driver.findElement('id', 'useragent');
+      return driver.execute(SCROLL_INTO_VIEW, [{'ELEMENT': (el.value + 1)}]).should.beRejected;
+    });
+
+    it('should be able to return multiple elements from javascript', async () => {
+      let res = await driver.execute(GET_ELEM_BY_TAGNAME);
+      expect(res).to.have.length.above(0);
+    });
+    it('should pass along non-element arguments', async () => {
+      let arg = 'non-element-argument';
+      (await driver.execute('var args = Array.prototype.slice.call(arguments, 0); return args[0];', [arg])).should.be.equal(arg);
+    });
+    it('should handle return values correctly', async () => {
+      let arg = ['one', 'two', 'three'];
+      (await driver.execute('var args = Array.prototype.slice.call(arguments, 0); return args;', arg)).should.eql(arg);
+    });
   });
 
-  it('should eval javascript', async () => {
-    (await driver.execute('return 1')).should.be.equal(1);
-  });
+  describe('asynchronous', function () {
+    it('should bubble up javascript errors', async () => {
+      await driver.execute(`'nan'--`).should.eventually.be.rejected;
+    });
 
-  it('should not be returning hardcoded results', async () => {
-    (await driver.execute('return 1+1')).should.be.equal(2);
-  });
+    it('should execute async javascript', async () => {
+      await driver.asyncScriptTimeout(1000);
+      (await driver.executeAsync(`arguments[arguments.length - 1](123);`)).should.be.equal(123);
+    });
 
-  it(`should return nothing when you don't explicitly return`, async () => {
-    expect(await driver.execute('1+1')).to.not.exist;
-  });
-
-  it('should execute code inside the web view', async () => {
-    (await driver.execute(GET_RIGHT_INNERHTML)).should.be.ok;
-    (await driver.execute(GET_WRONG_INNERHTML)).should.not.be.ok;
-  });
-
-  it('should convert selenium element arg to webview element', async () => {
-    let el = await driver.findElement('id', 'useragent');
-    await driver.execute(SCROLL_INTO_VIEW, [el]);
-  });
-
-  it('should catch stale or undefined element as arg', async () => {
-    let el = await driver.findElement('id', 'useragent');
-    return driver.execute(SCROLL_INTO_VIEW, [{'ELEMENT': (el.value + 1)}]).should.beRejected;
-  });
-
-  it('should be able to return multiple elements from javascript', async () => {
-    let res = await driver.execute(GET_ELEM_BY_TAGNAME);
-    expect(res).to.have.length.above(0);
-  });
-  it('should pass along non-element arguments', async () => {
-    let arg = 'non-element-argument';
-    (await driver.execute('var args = Array.prototype.slice.call(arguments, 0); return args[0];', [arg])).should.be.equal(arg);
-  });
-  it('should handle return values correctly', async () => {
-    let arg = ['one', 'two', 'three'];
-    (await driver.execute('var args = Array.prototype.slice.call(arguments, 0); return args;', arg)).should.eql(arg);
+    it('should timeout when callback is not invoked', async () => {
+      await driver.asyncScriptTimeout(1000);
+      await driver.executeAsync(`return 1 + 2`).should.eventually.be.rejected;
+    });
   });
 });
