--- test/e2e/safari/webview/window-title-specs.js
@@ -1,16 +1,0 @@
-/* globals expect */
-import desired from './desired';
-import setup from '../../setup-base';
-import { loadWebView } from '../../helpers/webview';
-
-describe("safari - webview - window title", function () {
-  const driver = setup(this, desired, {noReset: true}).driver;
-  beforeEach(async () => await loadWebView(desired, driver));
-
-  it('should return a valid title on web view', async () => {
-    (await driver.title()).should.include("I am a page title");
-
-    await driver.setContext('NATIVE_APP');
-    expect(async () => await driver.title()).to.throw;
-  });
-});
