--- test/e2e/uicatalog/window-specs.js
@@ -1,12 +1,0 @@
-import setup from "../setup-base";
-import desired from './desired';
-
-describe('uicatalog - contexts', function () {
-  let session = setup(this, desired);
-  let driver = session.driver;
-
-  it('getting contexts should do nothing when no webview open', async () => {
-    let contexts = await driver.getContexts();
-    contexts.should.have.length(1);
-  });
-});
